requirements.txt
python-dotenv>=1.0.1
pandas>=2.2.0
pillow>=10.2.0
numpy>=1.26.4